Add get_rhel_patch_groups.py and new variable A new variable  was added to the file variables.txt.example. It is used to define the path to Ansible's inventory file.
The script  imports this variable,
parses the rhel-patch-phases from the inventory and prints
them to STDOUT.

The script could be used standalone, to get a quick view on
which hosts are in which group. It is intended to be used in
a function called by  to generate a file from
the script's output that is going to be attachted to the email
generated by . Stay tuned.