New function  in
The two variables  and  were added to .
They are needed to run the script  and write it's output to the file defined
in .

A new function named  was added to  which writes the hosts in the
different rhel-patch-phases to the file specified in .

In case  is created successfully the file will be attached to the notification email.
In case the file does not exists the email will be send without attachment.